Debugged
Fixed annoying bug where objects and planes rendered in the wrong order. Created sorter class which takes in an array and outputs an index array sorted from max to min. Solved bug :)
Creating new mesh class to render terain maps (hopefully) tho will require perlin noise generator as well. For now mesh will take in a 2x2 double array of  heights and create a mesh. This is why fixing bug was important so planes render in the correct order within objects, as a mesh is a subclass of an object